impr: move sprite sheets resource to ryot_sprites
