+++ crates/ryot_core/src/content/sprite/sprite_sheet.rs
@@ -1,5 +1,4 @@
 use crate::prelude::SpriteLayout;
-use derive_more::{Deref, DerefMut};
 use glam::UVec2;
 use serde::{Deserialize, Serialize};
 
@@ -45,44 +44,5 @@ impl SpriteSheet {
         let width = self.layout.get_width(tile_size);
         let height = self.layout.get_height(tile_size);
         UVec2::new(width, height)
-    }
-}
-
-/// This is a collection of sprite sheets.
-/// It contains the sprite sheets and the sprite sheet config.
-/// The sprite sheet config is used to calculate the position and size of a sprite in the sprite
-/// sheet.
-#[derive(Debug, Default, Clone, Deref, DerefMut)]
-#[cfg_attr(feature = "bevy", derive(bevy_ecs::prelude::Resource))]
-pub struct SpriteSheets(Vec<SpriteSheet>);
-
-impl<T> From<&[T]> for SpriteSheets
-where
-    T: Into<Option<SpriteSheet>> + Clone,
-{
-    fn from(content: &[T]) -> Self {
-        let sprite_sheets = content
-            .iter()
-            .filter_map(|content_type| content_type.clone().into())
-            .collect::<Vec<_>>();
-
-        Self(sprite_sheets)
-    }
-}
-
-impl<T> From<Vec<T>> for SpriteSheets
-where
-    T: Into<Option<SpriteSheet>> + Clone,
-{
-    fn from(content: Vec<T>) -> Self {
-        content.as_slice().into()
-    }
-}
-
-impl SpriteSheets {
-    /// Returns the sprite sheet that contains the given sprite id.
-    /// Returns None if the sprite id is not in any of the sprite sheets.
-    pub fn get_by_sprite_id(&self, sprite_id: u32) -> Option<&SpriteSheet> {
-        self.iter().find(|sheet| sheet.has_sprite(sprite_id))
     }
 }
