+++ crates/ryot_sprites/src/tests/mod.rs
@@ -1,0 +1,1 @@
+mod sprite_sheets_test;
